a lot of work
- round to nearest instead of round to zero in SMS period calc
- implement arpeggio
- implement arp macro (kinda)
- polish the SMS platform
- correct the DIV_CMD_PITCH range to 1 semitone
- fix PSG volume in Genesis
- use a better register write strat in Genesis
- fix a bug caused by legacy code
- implement ECxx command
- implement EDxx command
- implement SN noise mode command
- vibrato table is now 64 positions long (instead of 60)

+++ src/engine/platform/abstract.cpp
@@ -1,6 +1,6 @@
 #include "../dispatch.h"
 
-void DivDispatch::acquire(short& l, short& r) {
+void DivDispatch::acquire(int& l, int& r) {
   l=0;
   r=0;
 }
